feat: brownie project support and dependencies that have dependencies (#642)
* refactor: separate project modules

* feat: able to compile brownie projects

* fix: use contracts_folder name in mapping

* test: test dependencies that have dependencies

* test: adjust test constants

* refactor: reorder classes

* chore: include more in all

* fix: undo accidental change
