+++ src/ape/managers/project/__init__.py
@@ -1,0 +1,13 @@
+from ape.managers.project.dependency import DependencyManager, GithubDependency, LocalDependency
+from ape.managers.project.manager import ProjectManager
+from ape.managers.project.types import ApeProject, BaseProject, BrownieProject
+
+__all__ = [
+    "ApeProject",
+    "BaseProject",
+    "BrownieProject",
+    "DependencyManager",
+    "GithubDependency",
+    "LocalDependency",
+    "ProjectManager",
+]
