+++ tests/integration/cli/projects/with-dependency/dependency_a/ape-config.yaml
@@ -1,0 +1,2 @@
+contracts_folder: sources
+name: __test_dependency_a__
