+++ tests/integration/cli/projects/with-dependency/dependency_a/sources/DependencyA.json
