+++ tests/integration/cli/projects/with-dependency/dependency_b/contracts/DependencyB.json
@@ -1,0 +1,3 @@
+[
+    {"name":"foo","type":"fallback", "stateMutability":"nonpayable"}
+]
