+++ tests/integration/cli/projects/with-dependency/dependency_b/dependency_b_a/ape-config.yaml
@@ -1,0 +1,2 @@
+contracts_folder: contracts
+name: __test_dependency_b_a__
