+++ tests/integration/cli/projects/with-dependency/dependency_b/dependency_b_a/contracts/DependencyBA.json
@@ -1,0 +1,3 @@
+[
+    {"name":"foo","type":"fallback", "stateMutability":"nonpayable"}
+]
